Scheduler works Debugging process related functions

+++ src/kernel/tests/test0.c
@@ -3,7 +3,8 @@
 #include "stdio.h"
 #include "assert.h"
 
-int64_t test0(void *arg)
+
+int test0(void *arg)
 {
 	(void)arg;
